[FIND-306] Updated idol-utils version to version 1.5.0.
This uses doubles rather than floats for some GetQueryTagValues response
objects.

Fixed tests.

[rev: ivo.miller]

+++ idol/src/main/java/com/hp/autonomy/searchcomponents/idol/parametricvalues/IdolParametricValuesService.java
@@ -206,9 +206,9 @@ public class IdolParametricValuesService implements ParametricValuesService<Idol
             for (final JAXBElement<?> element : valueElements) {
                 final String elementLocalName = element.getName().getLocalPart();
                 if (VALUE_MIN_NODE_NAME.equals(elementLocalName)) {
-                    minValue = (Float) element.getValue();
+                    minValue = (Double) element.getValue();
                 } else if (VALUE_MAX_NODE_NAME.equals(elementLocalName)) {
-                    maxValue = (Float) element.getValue();
+                    maxValue = (Double) element.getValue();
                 }
             }
